Add parallax support to all blocks
