+++ src/blocks/button-maxi/inspector.js
@@ -541,9 +541,7 @@ const Inspector = memo(
 													/>
 												),
 											},
-											...inspectorTabs.background({
-												props,
-											}),
+											...inspectorTabs.background(props),
 											{
 												label: __(
 													'Border',
