+++ src/blocks/column-maxi/inspector.js
@@ -194,7 +194,7 @@ const Inspector = props => {
 												</>
 											),
 										},
-										...inspectorTabs.background({ props }),
+										...inspectorTabs.background(props),
 										{
 											label: __('Border', 'maxi-blocks'),
 											disablePadding: true,
