+++ src/blocks/container-maxi/inspector.js
@@ -197,10 +197,7 @@ const Inspector = props => {
 												</>
 											),
 										},
-										...inspectorTabs.background({
-											props,
-											enableParallax: true,
-										}),
+										...inspectorTabs.background(props),
 										{
 											label: __('Border', 'maxi-blocks'),
 											disablePadding: true,
