+++ src/blocks/image-maxi/inspector.js
@@ -544,9 +544,7 @@ const Inspector = memo(
 													/>
 												),
 											},
-											...inspectorTabs.background({
-												props,
-											}),
+											...inspectorTabs.background(props),
 											{
 												label: __(
 													'Border',
