+++ src/blocks/number-counter-maxi/edit.js
@@ -25,11 +25,6 @@ import {
 import getStyles from './styles';
 
 /**
- * External dependencies
- */
-import { isEmpty } from 'lodash';
-
-/**
  * Icons
  */
 import { replay } from '../../icons';
@@ -234,17 +229,12 @@ class edit extends MaxiBlockComponent {
 		return getStyles(this.props.attributes);
 	}
 
-	get getCustomData() {
-		const { uniqueID } = this.props.attributes;
+	get getMaxiCustomData() {
+		const { attributes } = this.props;
 
 		return {
-			[uniqueID]: {
-				...{
-					...getGroupAttributes(
-						this.props.attributes,
-						'numberCounter'
-					),
-				},
+			...{
+				...getGroupAttributes(attributes, 'numberCounter'),
 			},
 		};
 	}
