+++ src/blocks/row-maxi/inspector.js
@@ -218,7 +218,7 @@ const Inspector = props => {
 												</>
 											),
 										},
-										...inspectorTabs.background({ props }),
+										...inspectorTabs.background(props),
 										{
 											label: __('Border', 'maxi-blocks'),
 											disablePadding: true,
