+++ src/blocks/svg-icon-maxi/inspector.js
@@ -360,7 +360,7 @@ const Inspector = props => {
 												/>
 											),
 										},
-										...inspectorTabs.background({ props }),
+										...inspectorTabs.background(props),
 										{
 											label: __('Border', 'maxi-blocks'),
 											disablePadding: true,
