+++ src/blocks/text-maxi/edit.js
@@ -42,20 +42,6 @@ class edit extends MaxiBlockComponent {
 		return getStyles(this.props.attributes);
 	}
 
-	get getCustomData() {
-		const { uniqueID } = this.props.attributes;
-
-		const motionStatus = !!this.props.attributes['motion-status'];
-
-		return {
-			[uniqueID]: {
-				...(motionStatus && {
-					...getGroupAttributes(this.props.attributes, ['motion']),
-				}),
-			},
-		};
-	}
-
 	render() {
 		const {
 			attributes,
