+++ src/blocks/text-maxi/inspector.js
@@ -382,9 +382,7 @@ const Inspector = memo(
 													/>
 												),
 											},
-											...inspectorTabs.background({
-												props,
-											}),
+											...inspectorTabs.background(props),
 											{
 												label: __(
 													'Border',
